requirements.txt
aiohttp==3.9.3
aiohttp-devtools==1.1.2
aiohttp-jinja2==1.6
aiosignal==1.3.1
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
click==8.1.7
devtools==0.12.2
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
pygments==2.17.2
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
typing-extensions==4.10.0
watchfiles==0.21.0
yarl==1.9.4
